requirements.txt
langgraph>=0.3
langgraph-checkpoint-sqlite>=2.0
langchain>=0.3
langchain-openai>=0.3
langchain-tavily>=0.1
tavily-python>=0.3
langchain-core>=0.3
firecrawl-py>=1.0
playwright>=1.40
pydantic>=2.0
aiosqlite
python-dotenv
fastapi
uvicorn
sse-starlette
streamlit
sseclient-py
